# backend/src/main/resources/api-docs.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "PLMS Parking System API",
    "description": "停车场管理系统API文档",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
    "/users/{userId}": {
      "put": {
        "tags": [
          "用户管理"
        ],
        "summary": "更新用户",
        "description": "更新一个用户",
        "operationId": "updateUser",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SysUser"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultBoolean"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "用户管理"
        ],
        "summary": "删除用户",
        "description": "删除一个用户",
        "operationId": "deleteUser",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultBoolean"
                }
              }
            }
          }
        }
      }
    },
    "/users/{userId}/reset-password": {
      "put": {
        "tags": [
          "用户管理"
        ],
        "summary": "重置密码",
        "description": "重置用户密码",
        "operationId": "resetPassword",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "newPassword",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultBoolean"
                }
              }
            }
          }
        }
      }
    },
    "/users/password": {
      "put": {
        "tags": [
          "用户管理"
        ],
        "summary": "修改密码",
        "description": "修改用户密码",
        "operationId": "updatePassword",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "oldPassword",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "newPassword",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultBoolean"
                }
              }
            }
          }
        }
      }
    },
...
    "/users": {
      "get": {
        "tags": [
          "用户管理"
        ],
        "summary": "获取用户列表",
        "description": "获取所有用户的列表",
        "operationId": "getUserList",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultListSysUser"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "用户管理"
        ],
        "summary": "创建用户",
        "description": "创建一个新的用户",
        "operationId": "createUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SysUser"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultBoolean"
                }
              }
            }
          }
        }
      }
    },
...
    "/users/page": {
      "get": {
        "tags": [
          "用户管理"
        ],
        "summary": "分页查询用户",
        "description": "分页查询用户列表",
        "operationId": "getUserPage",
        "parameters": [
          {
            "name": "current",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "default": 1
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "default": 10
            }
          },
          {
            "name": "username",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "roleId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultIPageSysUserVO"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "SysUser": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "description": "用户唯一标识"
          },
          "username": {
            "type": "string",
            "description": "登录用户名"
          },
          "password": {
            "type": "string",
            "description": "加密密码"
          },
          "roleId": {
            "type": "string",
            "description": "关联角色ID"
          },
          "department": {
            "type": "string",
            "description": "所属部门"
          },
          "status": {
            "type": "string",
            "description": "状态"
          }
        },
        "description": "存储操作人员信息"
      },
      "ResultBoolean": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "状态码（200=成功，其他=失败）",
            "format": "int32",
            "example": 200
          },
          "msg": {
            "type": "string",
            "description": "响应消息（成功/失败描述）",
            "example": "操作成功"
          },
          "data": {
            "type": "boolean",
            "description": "响应数据（成功时返回，失败时为null）"
          },
          "success": {
            "type": "boolean"
          }
        },
        "description": "统一响应结果"
      },
...
      "ResultListSysUser": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "状态码（200=成功，其他=失败）",
            "format": "int32",
            "example": 200
          },
          "msg": {
            "type": "string",
            "description": "响应消息（成功/失败描述）",
            "example": "操作成功"
          },
          "data": {
            "type": "array",
            "description": "响应数据（成功时返回，失败时为null）",
            "items": {
              "$ref": "#/components/schemas/SysUser"
            }
          },
          "success": {
            "type": "boolean"
          }
        },
        "description": "统一响应结果"
      },
      "IPageSysUserVO": {
        "type": "object",
        "properties": {
          "size": {
            "type": "integer",
            "format": "int64"
          },
          "records": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SysUserVO"
            }
          },
          "current": {
            "type": "integer",
            "format": "int64"
          },
          "pages": {
            "type": "integer",
            "format": "int64"
          },
          "total": {
            "type": "integer",
            "format": "int64"
          }
        },
        "description": "响应数据（成功时返回，失败时为null）"
      },
      "ResultIPageSysUserVO": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "状态码（200=成功，其他=失败）",
            "format": "int32",
            "example": 200
          },
          "msg": {
            "type": "string",
            "description": "响应消息（成功/失败描述）",
            "example": "操作成功"
          },
          "data": {
            "$ref": "#/components/schemas/IPageSysUserVO"
          },
          "success": {
            "type": "boolean"
          }
        },
        "description": "统一响应结果"
      },
      "SysUserVO": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "description": "用户ID"
          },
          "username": {
            "type": "string",
            "description": "用户名"
          },
          "roleId": {
            "type": "string",
            "description": "角色ID"
          },
          "roleName": {
            "type": "string",
            "description": "角色名称"
          },
          "department": {
            "type": "string",
            "description": "部门"
          },
          "status": {
            "type": "string",
            "description": "状态（01=启用，02=禁用）"
          },
          "statusDesc": {
            "type": "string",
            "description": "状态描述"
          }
        },
        "description": "用户详情VO"
      },
...
    }
  }
}